add new tls config from rpc client
