+++ src/main/scala/io/iohk/ethereum/crypto/EcKeyGen.scala
@@ -1,6 +1,6 @@
 package io.iohk.ethereum.crypto
 
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 
 /**
   * A simple tool to generate and ECDSA key pair. The key pair will be printed in the format:
