+++ src/main/scala/io/iohk/ethereum/security/SecureRandomBuilder.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.jsonrpc.security
+package io.iohk.ethereum.security
 
 import java.security.SecureRandom
