+++ src/test/scala/io/iohk/ethereum/domain/SignedTransactionSpec.scala
@@ -3,7 +3,7 @@ package io.iohk.ethereum.domain
 import io.iohk.ethereum.crypto
 import io.iohk.ethereum.crypto.generateKeyPair
 import io.iohk.ethereum.domain.SignedTransaction.FirstByteOfAddress
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import io.iohk.ethereum.vm.Generators
 import org.scalacheck.Arbitrary
 import org.bouncycastle.crypto.params.ECPublicKeyParameters
