+++ src/test/scala/io/iohk/ethereum/jsonrpc/NetServiceSpec.scala
@@ -6,7 +6,7 @@ import java.util.concurrent.atomic.AtomicReference
 import akka.actor.ActorSystem
 import akka.testkit.TestProbe
 import io.iohk.ethereum.jsonrpc.NetService._
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import io.iohk.ethereum.network.{Peer, PeerActor, PeerManagerActor}
 import io.iohk.ethereum.utils.{NodeStatus, ServerStatus}
 import io.iohk.ethereum.{NormalPatience, crypto}
