+++ src/test/scala/io/iohk/ethereum/network/AuthInitiateMessageSpec.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.network
 
 import akka.util.ByteString
 import io.iohk.ethereum.crypto._
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import io.iohk.ethereum.network.rlpx.{AuthHandshaker, AuthInitiateMessage}
 import io.iohk.ethereum.utils.ByteUtils
 import org.bouncycastle.crypto.generators.ECKeyPairGenerator
