+++ src/test/scala/io/iohk/ethereum/network/handshaker/EtcHandshakerSpec.scala
@@ -20,7 +20,7 @@ import io.iohk.ethereum.network.p2p.messages.WireProtocol.{Capability, Disconnec
 import io.iohk.ethereum.utils._
 import java.util.concurrent.atomic.AtomicReference
 
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
