+++ src/test/scala/io/iohk/ethereum/network/p2p/SecureChannelSetup.scala
@@ -5,7 +5,7 @@ import java.net.URI
 import akka.util.ByteString
 import io.iohk.ethereum.crypto
 import io.iohk.ethereum.crypto._
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import io.iohk.ethereum.network._
 import io.iohk.ethereum.network.rlpx.{AuthHandshakeSuccess, AuthHandshaker, Secrets}
 import org.bouncycastle.crypto.AsymmetricCipherKeyPair
