+++ src/test/scala/io/iohk/ethereum/network/p2p/messages/NewBlockSpec.scala
@@ -6,7 +6,7 @@ import io.iohk.ethereum.domain.{Block, BlockBody, BlockHeader, ChainWeight}
 import io.iohk.ethereum.network.p2p.messages.CommonMessages.NewBlock
 import org.bouncycastle.util.encoders.Hex
 import NewBlock._
-import io.iohk.ethereum.jsonrpc.security.SecureRandomBuilder
+import io.iohk.ethereum.security.SecureRandomBuilder
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import org.scalatest.funsuite.AnyFunSuite
