+++ src/universal/conf/faucet.conf
@@ -24,6 +24,22 @@ faucet {
   # Address of Ethereum node used to send the transaction
   rpc-address = "http://127.0.0.1:8546/"
 
+  # certificate of Ethereum node used to send the transaction when use HTTP(S)
+  certificate = null
+  #certificate {
+  # Path to the keystore storing the certificates (used only for https)
+  # null value indicates HTTPS is not being used
+  #  keystore-path = "tls/mantisCA.p12"
+
+  # Type of certificate keystore being used
+  # null value indicates HTTPS is not being used
+  #  keystore-type = "pkcs12"
+
+  # File with the password used for accessing the certificate keystore (used only for https)
+  # null value indicates HTTPS is not being used
+  #  password-file = "tls/password"
+  #}
+
   # How often can a single IP address send a request
   min-request-interval = 1.minute
 
@@ -80,6 +96,7 @@ mantis {
         # Listening port of JSON-RPC HTTP(S) endpoint
         port = 8099
 
+        # certificate when use JSON-RPC HTTP(S)
         certificate = null
         #certificate {
         # Path to the keystore storing the certificates (used only for https)
